+++ backend/routes2.js
@@ -67,7 +67,7 @@ app.get('/api/states', async (req, res) => {
         )
         SELECT Y.state_name AS "stateName",
                D.disease_name AS "diseaseName",
-               (W.yearly_cases_total::NUMERIC / NULLIF(Y.population, 0)) AS "perCapitaYearlyCases"
+               (W.yearly_cases_total::FLOAT / NULLIF(Y.population, 0)) AS "perCapitaYearlyCases"
         FROM yearly_cases W
         JOIN year_state_populations Y ON W.region_id = Y.region_id
         JOIN dim_disease D ON D.disease_id = W.disease_id
